requirements.txt
aiogram>=3.7
asyncpg>=0.29
pydantic-settings>=2.2
python-dotenv>=1.0
aiohttp>=3.9
gspread>=6.1
google-auth>=2.33